docker-compose/services/flask-restless-example/src/requirements.txt
Flask-SQLAlchemy
Flask-Restless
gunicorn
flask-cors
#ujson
#requests
